+++ src/containerapp/HISTORY.rst
@@ -7,6 +7,7 @@ Release History
 ++++++
 * Add 'az containerapp env update' to update managed environment properties
 * Add custom domains support to 'az containerapp env create' and 'az containerapp env update'
+* 'az containerapp compose create': Migrated from containerapp-compose extension
 
 0.3.11
 ++++++